use production ready connect-redis instead of in-memory
