+++ package.json
@@ -13,13 +13,14 @@
 	"license": "MIT",
 	"author": "codeshelldev",
 	"type": "commonjs",
-	"main": "app.js",
+	"main": "server.js",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
-		"start": "node ./src/app.js"
+		"start": "node ./src/server.js"
 	},
 	"dependencies": {
 		"axios": "^1.9.0",
+		"connect-redis": "^10.0.0",
 		"ejs": "^3.1.10",
 		"express": "^5.1.0",
 		"express-session": "^1.18.1",
